Fixes after some testing with a small test compiler.
The handle-language-extenion function was assuming terminals was always
going to be a single clause (which, admittedly it is in testing so far),
and was also dropping nonterminals that were not altered on the floor.
Extending languages now gather up all of the new terminal and
nonterminal clauses, removing terminals and nonterminal clauses as it
goes, and then rewrites any terminal or nonterminal clauses.