Simplified pattern language for experimental fork.
In the original nanopass-framework patterns could be in the following
forms:
id           - an identifier (reference to a meta-variable)
(maybe id)   - an optional identifier (reference to a meta-variable)
(x ellipsis) - a pattern followed by an ellipsis
(x ellipsis y ... . z) - a pattern, followed by an ellipsis, followed by
                         zero or more patterns, followed by pattern,
                         where the final pattern finishes an improper list.
(x . y)      - a pair of pattern and pattern
null         - the literal null
However, you might notice that the dotted patterns are a bit redundant,
namely (x ellipsis) is the same as (x ellipsis y ... . z) where y is the
empty list and z is null. So, why have a distinct form?  No reason.
Additionally, z could itself be a list (proper or improper) of patterns
modeled through the pair clause, so we are left with:
id           - an identifier (reference to a meta-variable)
(maybe id)   - an optional identifier (reference to a meta-variable)
(x ellipsis . z) - a pattern, followed by an ellipsis, followed by
                   a pattern where the final pattern finishes an
                   improper list.
(x . y)      - a pair of pattern and pattern
null         - the literal null